[script.tvmaze.scrobbler] 1.1.0
