+++ script.tvmaze.scrobbler/addon.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <addon id="script.tvmaze.scrobbler"
    name="TVmaze Scrobbler/Tracker"
-   version="1.0.1"
+   version="1.1.0"
    provider-name="Roman V.M.">
   <requires>
     <import addon="xbmc.python" version="2.26.0" />
@@ -38,7 +38,11 @@ What this addon does:
       <icon>resources/images/icon.png</icon>
       <fanart>resources/images/background.jpg</fanart>
     </assets>
-    <news>1.0.1:
+    <news>1.1.0:
+- Added periodic pulling of episode watched statuses from TVmaze.
+- Improved error logging.
+
+1.0.1:
 - Fixed crashes because of malformed JSON responses from TVmaze.
 - Fixed crashes because of invalid SSL certificates.</news>
   </extension>
